Add tests for ContractsPage
