+++ src/Client/Pages/ContractCard.razor
